requirements.txt
Flask
flask-cors
yt-dlp
pywebview
